week-01_local-llm-cli/requirements.txt
llama-cpp-python==0.2.60
textblob==0.17.1
click==8.1.7